Refactor folders name to vehicles and tickets
